+++ packages/google_ml_kit/example/lib/vision_detector_views/object_detector_view.dart
@@ -1,6 +1,11 @@
+import 'dart:io' as io;
+
 import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:google_mlkit_object_detection/google_mlkit_object_detection.dart';
+import 'package:path/path.dart';
+import 'package:path_provider/path_provider.dart';
 
 import 'camera_view.dart';
 import 'painters/object_detector_painter.dart';
@@ -11,23 +16,17 @@ class ObjectDetectorView extends StatefulWidget {
 }
 
 class _ObjectDetectorView extends State<ObjectDetectorView> {
-  final LocalModel _model = LocalModel('object_labeler.tflite');
   late ObjectDetector _objectDetector;
-  final bool _useLocalModel = false;
-  bool _canProcess = true;
+  bool _canProcess = false;
   bool _isBusy = false;
   CustomPaint? _customPaint;
   String? _text;
 
   @override
   void initState() {
-    _objectDetector = ObjectDetector(
-        options: _useLocalModel
-            ? CustomObjectDetectorOptions(_model,
-                multipleObjects: true, classifyObjects: true)
-            : ObjectDetectorOptions(
-                classifyObjects: true, multipleObjects: true));
     super.initState();
+
+    _initializeDetector();
   }
 
   @override
@@ -48,6 +47,39 @@ class _ObjectDetectorView extends State<ObjectDetectorView> {
       },
       initialDirection: CameraLensDirection.back,
     );
+  }
+
+  void _initializeDetector() async {
+    // uncomment next lines if you want to use the default model
+    // final options =
+    //     ObjectDetectorOptions(classifyObjects: true, multipleObjects: true);
+    // _objectDetector = ObjectDetector(options: options);
+
+    // uncomment next lines if you want to use a local model
+    // make sure to add tflite model to assets/ml
+    final path = 'assets/ml/object_labeler.tflite';
+    final modelPath = await _getModel(path);
+    final options = LocalObjectDetectorOptions(
+      modelPath: modelPath,
+      classifyObjects: true,
+      multipleObjects: true,
+    );
+    _objectDetector = ObjectDetector(options: options);
+
+    // uncomment next lines if you want to use a remote model
+    // make sure to add model to firebase
+    // final modelName = 'bird-classifier';
+    // final response =
+    //     await FirebaseObjectDetectorModelManager().downloadModel(modelName);
+    // print('Downloaded: $response');
+    // final options = FirebaseObjectDetectorOptions(
+    //   modelName: modelName,
+    //   classifyObjects: true,
+    //   multipleObjects: true,
+    // );
+    // _objectDetector = ObjectDetector(options: options);
+
+    _canProcess = true;
   }
 
   Future<void> processImage(InputImage inputImage) async {
@@ -79,5 +111,20 @@ class _ObjectDetectorView extends State<ObjectDetectorView> {
     if (mounted) {
       setState(() {});
     }
+  }
+
+  Future<String> _getModel(String assetPath) async {
+    if (io.Platform.isAndroid) {
+      return 'flutter_assets/$assetPath';
+    }
+    final path = '${(await getApplicationSupportDirectory()).path}/$assetPath';
+    await io.Directory(dirname(path)).create(recursive: true);
+    final file = io.File(path);
+    if (!await file.exists()) {
+      final byteData = await rootBundle.load(assetPath);
+      await file.writeAsBytes(byteData.buffer
+          .asUint8List(byteData.offsetInBytes, byteData.lengthInBytes));
+    }
+    return file.path;
   }
 }