+++ packages/google_mlkit_object_detection/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.0
 
 * Fix: return after closing detector in iOS.
+* Refactor [ObjectDetectorOptions] and subclasses.
+* Add example for loading local custom model.
 
 ## 0.1.0